+++ lib/metronome_method_channel.dart
@@ -34,6 +34,7 @@ class MethodChannelMetronome extends MetronomePlatform {
     bool enableTickCallback = false,
     int timeSignature = 4,
     int sampleRate = 44100,
+    bool manageAudioSession = true,
   }) async {
     if (mainPath == '') {
       throw Exception('Main path cannot be empty');
@@ -64,6 +65,7 @@ class MethodChannelMetronome extends MetronomePlatform {
         'enableTickCallback': enableTickCallback,
         'timeSignature': timeSignature,
         'sampleRate': sampleRate,
+        'manageAudioSession': manageAudioSession,
       });
     } catch (e) {
       if (kDebugMode) {